+++ __tests__/schema.test.ts
@@ -1,11 +1,8 @@
 import test from 'node:test';
-import assert, { strictEqual, throws } from 'node:assert';
-import { Durationable, Options, yamlPattern } from '../src/schema.ts';
+import { throws } from 'node:assert';
+import { Durationable, Options } from '../src/schema.ts';
 import { Temporal } from 'temporal-polyfill';
 import { durationEqual, optionsEqual } from './assert.ts';
-import { z } from 'zod';
-import { deepStrictEqual } from 'node:assert/strict';
-import { checkSync } from 'recheck';
 
 const defaultOptions = Object.freeze({
   apiUrl: 'https://api.github.com',
@@ -13,8 +10,8 @@ const defaultOptions = Object.freeze({
   attemptLimits: 1000,
   waitList: [],
   skipList: [],
-  warmupDelay: Temporal.Duration.from({ seconds: 10 }),
-  minimumInterval: Temporal.Duration.from({ seconds: 15 }),
+  warmupDelay: Temporal.Duration.from({ seconds: 1 }),
+  minimumInterval: Temporal.Duration.from({ seconds: 10 }),
   retryMethod: 'equal_intervals',
   shouldSkipSameWorkflow: false,
   isDryRun: false,
@@ -28,8 +25,8 @@ test('Options keep given values', () => {
     attemptLimits: 1000,
     waitList: [],
     skipList: [],
-    warmupDelay: Temporal.Duration.from({ seconds: 10 }),
-    minimumInterval: Temporal.Duration.from({ seconds: 15 }),
+    warmupDelay: Temporal.Duration.from({ seconds: 1 }),
+    minimumInterval: Temporal.Duration.from({ seconds: 10 }),
     retryMethod: 'equal_intervals',
     shouldSkipSameWorkflow: false,
     isDryRun: false,
@@ -82,10 +79,6 @@ test('Options accept all yaml extensions', () => {
       }],
     },
   );
-});
-
-test('regex option does not have higher ReDoS possibilities', () => {
-  strictEqual(checkSync(yamlPattern.source, '').status, 'safe');
 });
 
 test('It can start immediately. GH-994', () => {
@@ -148,49 +141,9 @@ test('Options reject invalid values', () => {
         ...defaultOptions,
         waitList: [{ workflowFile: 'ci.toml' }],
       }),
-    (err) => {
-      assert(err instanceof z.ZodError);
-      deepStrictEqual(err.issues, [
-        {
-          code: 'invalid_format',
-          format: 'regex',
-          message: 'Invalid string: must match pattern /\\.(yml|yaml)$/',
-          origin: 'string',
-          path: [
-            'waitList',
-            0,
-            'workflowFile',
-          ],
-          pattern: '/\\.(yml|yaml)$/',
-        },
-      ]);
-      return true;
-    },
-  );
-
-  throws(
-    () =>
-      Options.parse({
-        ...defaultOptions,
-        waitList: [{ workflowFile: 'ciyaml' }],
-      }),
-    (err) => {
-      assert(err instanceof z.ZodError);
-      deepStrictEqual(err.issues, [
-        {
-          code: 'invalid_format',
-          format: 'regex',
-          message: 'Invalid string: must match pattern /\\.(yml|yaml)$/',
-          origin: 'string',
-          path: [
-            'waitList',
-            0,
-            'workflowFile',
-          ],
-          pattern: '/\\.(yml|yaml)$/',
-        },
-      ]);
-      return true;
+    {
+      name: 'ZodError',
+      message: /Invalid string: must end with /,
     },
   );
 });
@@ -198,7 +151,10 @@ test('Options reject invalid values', () => {
 test('Durationable', async (t) => {
   await t.test('transformed to Temporal.Duration', (_t) => {
     durationEqual(Durationable.parse('PT1M42S'), Temporal.Duration.from({ seconds: 102 }));
-    durationEqual(Durationable.parse({ minutes: 1, seconds: 42 }), Temporal.Duration.from({ seconds: 102 }));
+    durationEqual(
+      Durationable.parse(Temporal.Duration.from({ minutes: 1, seconds: 42 })),
+      Temporal.Duration.from({ seconds: 102 }),
+    );
   });
 
   await t.test('it raises an error if given an invalid formats', (_t) => {
@@ -207,16 +163,6 @@ test('Durationable', async (t) => {
       {
         name: 'ZodError',
         message: /invalid_format/,
-      },
-    );
-  });
-
-  await t.test('it raises an error if given an unexpected keys', (_t) => {
-    throws(
-      () => Durationable.parse({ min: 5 }),
-      {
-        name: 'ZodError',
-        message: /unrecognized_key/,
       },
     );
   });
@@ -242,33 +188,16 @@ test('wait-list have startupGracePeriod', async (t) => {
     );
   });
 
-  await t.test('it raises an error if given an unexpected key', (_t) => {
+  await t.test('it raises an error if given an unexpected format', (_t) => {
     throws(
       () =>
         Options.parse({
           ...defaultOptions,
-          waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: { min: 5 } }],
+          waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: '5M' }],
         }),
       {
-        name: 'ZodError',
-        message: /Unrecognized key.+\bmin\b/,
-      },
-    );
-  });
-
-  await t.test('it raises a ZodError if given an unexpected keys', {
-    todo: "TODO: I don't know why using refine appears the native error",
-    skip: 'SKIP: To suppress noise',
-  }, (_t) => {
-    throws(
-      () =>
-        Options.parse({
-          ...defaultOptions,
-          waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: { min: 5 } }],
-        }),
-      {
-        name: 'ZodError',
-        message: /unrecognized_key/,
+        name: 'RangeError',
+        message: /Cannot parse:.+\b5M\b/,
       },
     );
   });
@@ -298,7 +227,7 @@ test('wait-list have startupGracePeriod', async (t) => {
         Options.parse({
           ...defaultOptions,
           warmupDelay: Temporal.Duration.from({ seconds: 41 }),
-          waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: { seconds: 40 } }],
+          waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: 'PT40S' }],
         }),
       {
         name: 'ZodError',
@@ -312,7 +241,7 @@ test('wait-list have startupGracePeriod', async (t) => {
       Options.parse({
         ...defaultOptions,
         warmupDelay: Temporal.Duration.from({ seconds: 42 }),
-        waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: { seconds: 10 } }],
+        waitList: [{ workflowFile: 'ci.yml', startupGracePeriod: 'PT10S' }],
       }),
       {
         ...defaultOptions,